Restructure deploy logic for env for consistency
