+++ internal/pkg/deploy/cloudformation/cloudformation_test.go
@@ -9,8 +9,7 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/archer"
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/deploy"
+	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/deploy/cloudformation/stack"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
@@ -404,161 +403,6 @@ func TestWaitForStackCreation(t *testing.T) {
 	}
 }
 
-func TestStreamEnvironmentCreation(t *testing.T) {
-	testCases := map[string]struct {
-		in *deploy.CreateEnvironmentInput
-
-		mockWaitUntilStackCreateComplete func(t *testing.T, in *cloudformation.DescribeStacksInput) error
-		mockDescribeStacks               func(t *testing.T, in *cloudformation.DescribeStacksInput) (*cloudformation.DescribeStacksOutput, error)
-		mockDescribeStackEvents          func(t *testing.T, in *cloudformation.DescribeStackEventsInput) (*cloudformation.DescribeStackEventsOutput, error)
-
-		wantedEvents []deploy.ResourceEvent
-		wantedResult deploy.CreateEnvironmentResponse
-	}{
-		"error while creating stack": {
-			in: &deploy.CreateEnvironmentInput{
-				Project: "phonetool",
-				Name:    "test",
-			},
-
-			mockWaitUntilStackCreateComplete: func(t *testing.T, in *cloudformation.DescribeStacksInput) error {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return errors.New("wait until error")
-			},
-			mockDescribeStackEvents: func(t *testing.T, in *cloudformation.DescribeStackEventsInput) (*cloudformation.DescribeStackEventsOutput, error) {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return &cloudformation.DescribeStackEventsOutput{
-					StackEvents: []*cloudformation.StackEvent{
-						{
-							LogicalResourceId:    aws.String("vpc"),
-							ResourceType:         aws.String("AWS::EC2::VPC"),
-							ResourceStatus:       aws.String(cloudformation.ResourceStatusCreateInProgress),
-							ResourceStatusReason: aws.String("create initiated"),
-						},
-					},
-				}, nil
-			},
-
-			wantedEvents: []deploy.ResourceEvent{
-				{
-					Resource: deploy.Resource{
-						LogicalName: "vpc",
-						Type:        "AWS::EC2::VPC",
-					},
-					Status:       cloudformation.ResourceStatusCreateInProgress,
-					StatusReason: "create initiated",
-				},
-			},
-			wantedResult: deploy.CreateEnvironmentResponse{
-				Env: nil,
-				Err: errors.New("failed to create stack phonetool-test: wait until error"),
-			},
-		},
-		"swallows error while describing stack": {
-			in: &deploy.CreateEnvironmentInput{
-				Project: "phonetool",
-				Name:    "test",
-			},
-
-			mockWaitUntilStackCreateComplete: func(t *testing.T, in *cloudformation.DescribeStacksInput) error {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return errors.New("wait until error")
-			},
-			mockDescribeStackEvents: func(t *testing.T, in *cloudformation.DescribeStackEventsInput) (*cloudformation.DescribeStackEventsOutput, error) {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return nil, errors.New("describe stack events error")
-			},
-
-			wantedEvents: nil,
-			wantedResult: deploy.CreateEnvironmentResponse{
-				Env: nil,
-				Err: errors.New("failed to create stack phonetool-test: wait until error"),
-			},
-		},
-		"sends an environment on success": {
-			in: &deploy.CreateEnvironmentInput{
-				Project: "phonetool",
-				Name:    "test",
-			},
-
-			mockWaitUntilStackCreateComplete: func(t *testing.T, in *cloudformation.DescribeStacksInput) error {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return nil
-			},
-			mockDescribeStacks: func(t *testing.T, in *cloudformation.DescribeStacksInput) (*cloudformation.DescribeStacksOutput, error) {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return &cloudformation.DescribeStacksOutput{
-					Stacks: []*cloudformation.Stack{
-						{
-							StackId: aws.String(fmt.Sprintf("arn:aws:cloudformation:eu-west-3:902697171733:stack/%s", "phonetool-test")),
-						},
-					},
-				}, nil
-			},
-			mockDescribeStackEvents: func(t *testing.T, in *cloudformation.DescribeStackEventsInput) (*cloudformation.DescribeStackEventsOutput, error) {
-				require.Equal(t, "phonetool-test", *in.StackName, "stack names should be equal to each other")
-				return &cloudformation.DescribeStackEventsOutput{
-					StackEvents: []*cloudformation.StackEvent{
-						{
-							LogicalResourceId:    aws.String("vpc"),
-							ResourceType:         aws.String("AWS::EC2::VPC"),
-							ResourceStatus:       aws.String(cloudformation.ResourceStatusCreateInProgress),
-							ResourceStatusReason: aws.String("create initiated"),
-						},
-					},
-				}, nil
-			},
-
-			wantedEvents: []deploy.ResourceEvent{
-				{
-					Resource: deploy.Resource{
-						LogicalName: "vpc",
-						Type:        "AWS::EC2::VPC",
-					},
-					Status:       cloudformation.ResourceStatusCreateInProgress,
-					StatusReason: "create initiated",
-				},
-			},
-			wantedResult: deploy.CreateEnvironmentResponse{
-				Env: &archer.Environment{
-					Project:   "phonetool",
-					Name:      "test",
-					Region:    "eu-west-3",
-					AccountID: "902697171733",
-				},
-				Err: nil,
-			},
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			cf := CloudFormation{
-				client: &mockCloudFormation{
-					t:                                t,
-					mockDescribeStackEvents:          tc.mockDescribeStackEvents,
-					mockDescribeStacks:               tc.mockDescribeStacks,
-					mockWaitUntilStackCreateComplete: tc.mockWaitUntilStackCreateComplete,
-				},
-				box: emptyEnvBox(),
-			}
-
-			// WHEN
-			events, resp := cf.StreamEnvironmentCreation(tc.in)
-
-			// THEN
-			require.Equal(t, tc.wantedEvents, <-events)
-			got := <-resp
-			if tc.wantedResult.Err != nil {
-				require.EqualError(t, got.Err, tc.wantedResult.Err.Error(), "expected %v got %v", tc.wantedResult.Err, got.Err)
-			} else {
-				require.Equal(t, tc.wantedResult, got)
-			}
-		})
-	}
-}
-
 func getMockWaitStackCreateCFClient(t *testing.T, stackName string, shouldThrowError, shouldReturnEmptyStacks bool) CloudFormation {
 	return CloudFormation{
 		client: &mockCloudFormation{
@@ -586,7 +430,7 @@ func getMockWaitStackCreateCFClient(t *testing.T, stackName string, shouldThrowE
 				}, nil
 			},
 		},
-		box: emptyEnvBox(),
+		box: emptyBox(),
 	}
 }
 
@@ -614,7 +458,7 @@ func emptyBox() packd.Box {
 func boxWithTemplateFile() packd.Box {
 	box := packd.NewMemoryBox()
 
-	box.AddString(envTemplatePath, mockTemplate)
+	box.AddString(stack.EnvTemplatePath, mockTemplate)
 
 	return box
 }
